Обновлена логика iebus адаптера

+++ include/iebus/types.hpp
@@ -29,7 +29,7 @@ using Bit     = std::uint8_t;
 using Pin     = std::uint8_t;
 using Byte    = std::uint8_t;
 using Data    = std::uint16_t;
-using Size    = std::uint8_t;
+using Size    = std::uint16_t;
 using Bytes   = std::array<Byte, MAX_MESSAGE_SIZE>;
 using Address = std::uint16_t;
